[CALCULATOR-26] Вынести классы по отдельным файлам (разбить на модули)
