+++ src/index.html
@@ -3,7 +3,7 @@
     <head>
         <meta charset="UTF-8">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <link rel="stylesheet" href="css/style.min.css">
+        <link rel="stylesheet" href="css/style.css">
         <title>Calculator</title>
     </head>
     <body>
@@ -11,6 +11,6 @@
             <div class="calculator"></div>
         </div>
 
-        <script src="js/script.min.js"></script>
+        <script type="module" src="js/script.js"></script>
     </body>
 </html>
