+++ src/js/button_childs.js
@@ -1,0 +1,33 @@
+import {Button} from "./dom-render-elemen_childs.js";
+
+class ButtonNumber extends Button{
+    constructor(calculatorHTMLClass) {
+        super(calculatorHTMLClass);
+    }
+}
+
+class ButtonCleanup extends Button{
+    constructor(calculatorHTMLClass) {
+        super(calculatorHTMLClass);
+    }
+}
+
+class ButtonComplexOperation extends Button{
+    constructor(calculatorHTMLClass) {
+        super(calculatorHTMLClass);
+    }
+}
+
+class ButtonBasicOperation extends Button{
+    constructor(calculatorHTMLClass) {
+        super(calculatorHTMLClass);
+    }
+}
+
+class ButtonEqual extends Button{
+    constructor(calculatorHTMLClass) {
+        super(calculatorHTMLClass);
+    }
+}
+
+export {ButtonNumber, ButtonCleanup, ButtonComplexOperation, ButtonBasicOperation, ButtonEqual}
