+++ src/js/dom-render-element.js
@@ -1,0 +1,49 @@
+import {ELEMENTS_PROPERTY} from "./global_elements.js";
+
+export default class DomRendererElement {
+    constructor(calculatorHTMLClass) {
+        this.calculatorHTMLClass = calculatorHTMLClass;
+
+        this.setDefaultCalculatorHTMLClass();
+    }
+
+    setDefaultCalculatorHTMLClass() {
+        if (!this.calculatorHTMLClass) {
+            this.calculatorHTMLClass = ELEMENTS_PROPERTY.ROOT_FOR_MAIN;
+        }
+    }
+
+    render(params) {
+        const {
+            tagName,
+            classNames = [],
+            textContent,
+            datasetText,
+            datasetType,
+            switchOfButtonsClickability
+        } = params;
+
+        const element = document.createElement(tagName);
+
+        if (classNames) {
+            classNames.forEach(item => {
+                element.classList.add(`${item}`);
+            });
+        }
+        if (textContent) {
+            element.textContent = textContent;
+        }
+        if (datasetText) {
+            element.dataset.text = datasetText;
+        }
+        if (datasetType) {
+            element.dataset.type = datasetType;
+        }
+        if (switchOfButtonsClickability) {
+            element.style.color = "gray";
+            element.style.cursor = "inherit";
+        }
+
+        return element;
+    }
+}
